refactor: 💡 SocialAuthenticationResponseMapper mapper패키지로 이동
✅ Closes: #113
