+++ src/main/java/com/postsquad/scoup/web/auth/service/GitHubOAuth.java
@@ -4,7 +4,7 @@ import com.postsquad.scoup.web.auth.controller.request.TokenRequest;
 import com.postsquad.scoup.web.auth.controller.response.TokenResponse;
 import com.postsquad.scoup.web.auth.controller.response.GitHubUserResponse;
 import com.postsquad.scoup.web.auth.controller.response.SocialAuthenticationResponse;
-import com.postsquad.scoup.web.auth.controller.response.SocialAuthenticationResponseMapper;
+import com.postsquad.scoup.web.auth.mapper.SocialAuthenticationResponseMapper;
 import com.postsquad.scoup.web.auth.exception.TokenNotFoundException;
 import com.postsquad.scoup.web.auth.exception.OAuthRequestNotValidException;
 import com.postsquad.scoup.web.auth.exception.OAuthUserNotFoundException;
